Fix CI doc build failure (not reproducible locally) by changing DUST merkle index type to u64

+++ indexer-api/src/domain/dust.rs
@@ -187,7 +187,7 @@ pub struct DustCommitmentInfo {
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct DustCommitmentMerkleUpdate {
     /// Tree index.
-    pub index: u32,
+    pub index: u64,
 
     /// Collapsed update data.
     pub collapsed_update: DustMerkleUpdate,
